requirements.txt
Django==3.1
psycopg2-binary==2.8.6
Pillow==8.1
djangorestframework==3.12.4
python-decouple==3.4
django-ckeditor==6.0.0
pytils
django-filter